pyproject.toml
[tool.poetry]
name = "telegram-rag-bot"
version = "0.1.0"
description = ""
authors = ["Roman <roman.nieronov@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.8"
langchain = "^0.1.7"
beautifulsoup4 = "^4.12.3"
jinja2 = "^3.1.3"
openai = "^1.12.0"
pypdf = "^4.0.1"
langchain-community = "^0.0.20"
langchain-openai = "^0.0.6"
langchainhub = "^0.1.14"
chromadb = "^0.4.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
